+++ src/linter/LintManager.ts
@@ -9,6 +9,7 @@ import SlangLinter from './SlangLinter';
 import XvlogLinter from './XvlogLinter';
 import VeribleVerilogLintLinter from './VeribleVerilogLintLinter';
 import LinterDiagnosticManager from './LinterDiagnosticManager';
+import LintRunManager from './LintRunManager';
 
 export default class LintManager {
   private subscriptions: vscode.Disposable[];
@@ -16,6 +17,7 @@ export default class LintManager {
   private linter: BaseLinter | null;
   private diagnosticCollection: vscode.DiagnosticCollection;
   private diagnosticManager: LinterDiagnosticManager;
+  private runManager: LintRunManager;
   private readonly logger = getExtensionLogger('Linter', 'Manager');
 
   constructor() {
@@ -23,6 +25,7 @@ export default class LintManager {
     this.linter = null;
     this.diagnosticCollection = vscode.languages.createDiagnosticCollection('verilog-lint');
     this.diagnosticManager = new LinterDiagnosticManager(this.diagnosticCollection);
+    this.runManager = new LintRunManager();
     vscode.workspace.onDidOpenTextDocument(this.lint, this, this.subscriptions);
     vscode.workspace.onDidSaveTextDocument(this.lint, this, this.subscriptions);
     vscode.workspace.onDidCloseTextDocument(this.removeFileDiagnostics, this, this.subscriptions);
@@ -38,17 +41,17 @@ export default class LintManager {
   getLinterFromString(name: string): BaseLinter | null {
     switch (name) {
       case 'iverilog':
-        return new IcarusLinter(this.diagnosticManager);
+        return new IcarusLinter(this.diagnosticManager, this.runManager);
       case 'xvlog':
-        return new XvlogLinter(this.diagnosticManager);
+        return new XvlogLinter(this.diagnosticManager, this.runManager);
       case 'modelsim':
-        return new ModelsimLinter(this.diagnosticManager);
+        return new ModelsimLinter(this.diagnosticManager, this.runManager);
       case 'verilator':
-        return new VerilatorLinter(this.diagnosticManager);
+        return new VerilatorLinter(this.diagnosticManager, this.runManager);
       case 'slang':
-        return new SlangLinter(this.diagnosticManager);
+        return new SlangLinter(this.diagnosticManager, this.runManager);
       case 'verible-verilog-lint':
-        return new VeribleVerilogLintLinter(this.diagnosticManager);
+        return new VeribleVerilogLintLinter(this.diagnosticManager, this.runManager);
       default:
         return null;
     }
@@ -66,6 +69,7 @@ export default class LintManager {
 
     if (linterName === undefined) {
       if (previousLinter !== null) {
+        this.runManager.cancelSource(previousLinter.name);
         this.diagnosticManager.clearSource(previousLinter.name);
         this.linter = null;
       }
@@ -76,6 +80,7 @@ export default class LintManager {
     this.linter = this.getLinterFromString(linterName);
     if (this.linter === null) {
       if (previousLinter !== null) {
+        this.runManager.cancelSource(previousLinter.name);
         this.diagnosticManager.clearSource(previousLinter.name);
       }
       this.logger.warn("Invalid linter name", { linter: linterName });
@@ -83,6 +88,7 @@ export default class LintManager {
     }
 
     if (previousLinter !== null) {
+      this.runManager.cancelSource(previousLinter.name);
       this.diagnosticManager.clearSource(previousLinter.name);
     }
     this.logger.info("Linter configured", { linter: this.linter.name });
@@ -95,7 +101,7 @@ export default class LintManager {
     switch (doc.languageId) {
       case 'verilog':
       case 'systemverilog':
-        this.linter.startLint(doc);
+        void this.linter.startLint(doc);
         break;
       default:
         break;
@@ -103,10 +109,12 @@ export default class LintManager {
   }
 
   removeFileDiagnostics(doc: vscode.TextDocument) {
+    this.runManager.cancelOwner(doc.uri);
     this.diagnosticManager.clearTargetUri(doc.uri);
   }
 
   dispose(): void {
+    this.runManager.cancelAll();
     for (const subscription of this.subscriptions) {
       subscription.dispose();
     }
@@ -177,7 +185,7 @@ export default class LintManager {
         this.logger.info("Running linter", { linter: linter.name });
 
         linter.removeFileDiagnostics(editor.document);
-        linter.startLint(editor.document);
+        await linter.startLint(editor.document);
       }
     );
   }